+++ CHANGELOG.md
@@ -14,6 +14,12 @@ and this project tries to adhere to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - changed naming of vectorToList to cVectorToPython
+- Updated getspice module to use urllib3 for OpenSSL library versions older
+  than OpenSSL 1.0.1g.
+- getspice module provides now a class that handles the downloading and
+  unpacking of N066 CSPICE distribution.
+- Updated setup to pack the CSPICE installation code into a class that
+  extends the setuptools.command.install command.
 
 ## [2.0.0] - 2017-06-09
 ### Added